+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 🔥 *Features*
 * Secrets can now be used inside workflow functions
-
+* `sdk.secrets.get("name")` will now use passport-based authorization if `ORQUESTRA_PASSPORT_FILE` environment variable is set. Otherwise, passing a valid `config_name="..."` is required. Previously, "passport auth" was only used inside QE tasks; any other execution context required "config auth" regardless of environment variables.
 
 👩‍🔬 *Experimental*